# src/NovaLab/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NovaLab API",
    "description": "An ASP.NET Core Web API for managing your streams",
    "version": "v1"
  },
  "paths": {
...
    "/api/twitch/followers/follower-goal": {
      "get": {
        "tags": [
          "TwitchFollowerGoal"
        ],
        "operationId": "GetFollowerGoals",
        "parameters": [
          {
            "name": "user-id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "goal-ids",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FollowerGoalDtoApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowerGoalDtoApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowerGoalDtoApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TwitchFollowerGoal"
        ],
        "operationId": "PostFollowerGoal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostFollowerGoalDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostFollowerGoalDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostFollowerGoalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FollowerGoalDtoApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowerGoalDtoApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowerGoalDtoApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/twitch/followers/new-follower": {
      "get": {
        "tags": [
          "TwitchNewFollower"
        ],
        "operationId": "GetNewFollowers",
        "parameters": [
          {
            "name": "user-id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "goal-id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from-date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewFollowerDtoApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewFollowerDtoApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewFollowerDtoApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TwitchNewFollower"
        ],
        "operationId": "PostNewFollower",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostNewFollowerDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PostNewFollowerDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PostNewFollowerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NewFollowerDtoApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewFollowerDtoApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewFollowerDtoApiResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FollowerGoalDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "twitchBroadcasterId": {
            "type": "string",
            "nullable": true
          },
          "goalId": {
            "type": "string",
            "format": "uuid"
          },
          "dailyGoalAmount": {
            "type": "integer",
            "format": "int32"
          },
          "divider": {
            "type": "string",
            "nullable": true
          },
          "customCssStyling": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FollowerGoalDtoApiResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FollowerGoalDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HttpStatusCode": {
        "enum": [
          100,
          101,
          102,
          103,
          200,
          201,
          202,
          203,
          204,
          205,
          206,
          207,
          208,
          226,
          300,
          301,
          302,
          303,
          304,
          305,
          306,
          307,
          308,
          400,
          401,
          402,
          403,
          404,
          405,
          406,
          407,
          408,
          409,
          410,
          411,
          412,
          413,
          414,
          415,
          416,
          417,
          421,
          422,
          423,
          424,
          426,
          428,
          429,
          431,
          451,
          500,
          501,
          502,
          503,
          504,
          505,
          506,
          507,
          508,
          510,
          511
        ],
        "type": "integer",
        "format": "int32"
      },
      "NewFollowerDto": {
        "type": "object",
        "properties": {
          "newFollowerId": {
            "type": "string",
            "format": "uuid"
          },
          "followerGoalId": {
            "type": "string",
            "format": "uuid"
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "followerTwitchUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NewFollowerDtoApiResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NewFollowerDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PostFollowerGoalDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "dailyGoal": {
            "type": "integer",
            "format": "int32"
          },
          "divider": {
            "type": "string",
            "nullable": true
          },
          "customCssStyling": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PostNewFollowerDto": {
        "type": "object",
        "properties": {
          "twitchBoardcasterId": {
            "type": "string",
            "nullable": true
          },
          "followerTwitchUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}